Back to aether stable version.

+++ src/main/java/bee/definition/MavenVersionsMetadataGeneratorFactory.java
@@ -9,11 +9,11 @@
  */
 package bee.definition;
 
-import org.eclipse.aether.RepositorySystemSession;
-import org.eclipse.aether.deployment.DeployRequest;
-import org.eclipse.aether.impl.MetadataGenerator;
-import org.eclipse.aether.impl.MetadataGeneratorFactory;
-import org.eclipse.aether.installation.InstallRequest;
+import org.sonatype.aether.RepositorySystemSession;
+import org.sonatype.aether.deployment.DeployRequest;
+import org.sonatype.aether.impl.MetadataGenerator;
+import org.sonatype.aether.impl.MetadataGeneratorFactory;
+import org.sonatype.aether.installation.InstallRequest;
 
 /**
  * @version 2012/04/10 14:04:01
@@ -37,7 +37,7 @@ class MavenVersionsMetadataGeneratorFactory implements MetadataGeneratorFactory 
     /**
      * {@inheritDoc}
      */
-    public float getPriority() {
+    public int getPriority() {
         return 5;
     }
 }
